Move link context into @spark-web/link package
