+++ packages/button/package.json
@@ -12,8 +12,8 @@
     "@emotion/css": "^11.7.1",
     "@spark-web/a11y": "^0.0.0",
     "@spark-web/box": "^0.0.0",
-    "@spark-web/core": "^0.0.0",
     "@spark-web/icon": "^0.0.0",
+    "@spark-web/link": "^0.0.0",
     "@spark-web/text": "^0.0.0",
     "@spark-web/theme": "^0.0.0",
     "@spark-web/utils-spark": "^0.0.0",
