+++ packages/core/package.json
@@ -17,7 +17,6 @@
     "@spark-web/theme": "^0.0.0",
     "@spark-web/utils-ts": "^0.0.0",
     "assert": "1.5.0",
-    "dedent": "^0.7.0",
     "react": "^17.0.2"
   },
   "engines": {
