+++ packages/core/src/index.ts
@@ -3,10 +3,8 @@ export {
   FontStylesheets,
   InterStylesheet,
 } from './fonts';
-export { makeLinkComponent, useLinkComponent } from './linkContext';
 export { SparkProvider } from './SparkProvider';
 
 // types
 
-export type { LinkComponent, LinkComponentProps } from './linkContext';
 export type { SparkProviderProps } from './SparkProvider';
