+++ packages/link/package.json
@@ -10,8 +10,8 @@
   "dependencies": {
     "@babel/runtime": "^7.14.6",
     "@spark-web/box": "^0.0.0",
-    "@spark-web/core": "^0.0.0",
     "@spark-web/utils-ts": "^0.0.0",
+    "dedent": "^0.7.0",
     "react": "^17.0.2"
   },
   "engines": {
