+++ packages/link/src/Link.tsx
@@ -1,7 +1,8 @@
 import { Box } from '@spark-web/box';
-import type { LinkComponentProps } from '@spark-web/core';
-import { useLinkComponent } from '@spark-web/core';
 import { forwardRefWithAs } from '@spark-web/utils-ts';
+
+import type { LinkComponentProps } from './linkContext';
+import { useLinkComponent } from './linkContext';
 
 export type LinkProps = LinkComponentProps;
