+++ packages/link/src/index.ts
@@ -1,5 +1,12 @@
 export { Link } from './Link';
+export {
+  DefaultLinkComponent,
+  LinkComponentContext,
+  makeLinkComponent,
+  useLinkComponent,
+} from './linkContext';
 
 // types
 
 export type { LinkProps } from './Link';
+export type { LinkComponent, LinkComponentProps } from './linkContext';
