+++ packages/nav-link/package.json
@@ -8,7 +8,7 @@
     "@babel/runtime": "^7.14.6",
     "@emotion/css": "^11.7.1",
     "@spark-web/box": "0.0.0",
-    "@spark-web/core": "0.0.0",
+    "@spark-web/link": "^0.0.0",
     "@spark-web/text": "0.0.0",
     "@spark-web/theme": "0.0.0",
     "react": "^17.0.2"
