+++ packages/next-utils/package.json
@@ -9,7 +9,7 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.14.6",
-    "@spark-web/core": "0.0.0",
+    "@spark-web/link": "^0.0.0",
     "@spark-web/ssr": "0.0.0",
     "next": "^12.0.7",
     "react": "^17.0.2"
