+++ packages/text-link/src/TextLink.tsx
@@ -1,7 +1,7 @@
 import { css } from '@emotion/css';
 import { Box } from '@spark-web/box';
-import type { LinkComponentProps } from '@spark-web/core';
-import { useLinkComponent } from '@spark-web/core';
+import type { LinkComponentProps } from '@spark-web/link';
+import { useLinkComponent } from '@spark-web/link';
 import type { DataAttributeMap } from '@spark-web/utils-spark';
 import { buildDataAttributes } from '@spark-web/utils-spark';
 import { forwardRefWithAs } from '@spark-web/utils-ts';
